Cleanup wrapper decoding code
- Split into separate files for the tabular and object style data decoding
- Accept IEnumerable instead of String
- Inline operations from nested loops and add return early conditions
- Replace Custom Attribute with DataMemberAttribute
