+++ GithubPackageUpdater/GithubPackageUpdater.csproj
@@ -4,6 +4,7 @@
     <TargetPlatformMinVersion>10.0.17763.0</TargetPlatformMinVersion>
     <RootNamespace>GithubPackageUpdater</RootNamespace>
     <Nullable>warnings</Nullable>
+    <EnableNETAnalyzers>true</EnableNETAnalyzers>
   </PropertyGroup>
 
   <ItemGroup>
