+++ WingetHelper/Commands/WingetCommand.cs
@@ -28,7 +28,7 @@ namespace WingetHelper.Commands
 
         private Action<WingetProcessState> _progressMonitor;
         private Action<string> _outputListener;
-        private Func<List<string>, TResult> _resultDecoder;
+        private Func<IEnumerable<string>, TResult> _resultDecoder;
 
         internal WingetCommand(params string[] arguments)
         {
@@ -65,7 +65,7 @@ namespace WingetHelper.Commands
             return this;
         }
 
-        internal WingetCommand<TResult> ConfigureResultDecoder(Func<List<string>, TResult> resultDecoder)
+        internal WingetCommand<TResult> ConfigureResultDecoder(Func<IEnumerable<string>, TResult> resultDecoder)
         {
             _resultDecoder = resultDecoder;
             return this;
