+++ WingetHelper/WingetHelper.csproj
@@ -7,6 +7,7 @@
     <PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
     <RepositoryType>git</RepositoryType>
     <RepositoryUrl>https://github.com/zsolt3991/WingetGUIInstaller</RepositoryUrl>
+    <EnableNETAnalyzers>true</EnableNETAnalyzers>
     <Nullable>warnings</Nullable>
   </PropertyGroup>
   <ItemGroup>
